+++ semgrep-core/bin/Main.ml
@@ -9,7 +9,7 @@
  *)
 open Common
 
-module Flag = Flag_parsing
+module Flag = Flag_semgrep
 module PI = Parse_info
 module S = Scope_code
 module E = Error_code
@@ -55,11 +55,13 @@ module J = Json_type
 let env_debug = "SEMGREP_CORE_DEBUG"
 let env_profile = "SEMGREP_CORE_PROFILE"
 
+(* see also verbose/... flags in Flag_semgrep.ml *)
+(* to test things *)
+let test = ref false
+
 (*s: constant [[Main_semgrep_core.verbose]] *)
-let verbose = ref false
 (*e: constant [[Main_semgrep_core.verbose]] *)
 (*s: constant [[Main_semgrep_core.debug]] *)
-let debug = ref false
 (*e: constant [[Main_semgrep_core.debug]] *)
 let profile = ref false
 (*s: constant [[Main_semgrep_core.error_recovery]] *)
@@ -96,6 +98,7 @@ let lang = ref "unset"
 (*e: constant [[Main_semgrep_core.lang]] *)
 
 (* similar to grep options (see man grep) *)
+(* TODO: can remove now that handled by semgrep-python? *)
 (*s: constant [[Main_semgrep_core.excludes]] *)
 let excludes = ref []
 (*e: constant [[Main_semgrep_core.excludes]] *)
@@ -136,9 +139,13 @@ let supported_langs: string = String.concat ", " keys
 let ncores = ref 1
 (*e: constant [[Main_semgrep_core.ncores]] *)
 
+(* path to cache (given by semgrep-python) *)
 let use_parsing_cache = ref ""
+(* take the list of files in a file (given by semgrep-python) *)
 let target_file = ref ""
-let timeout = ref 0
+
+let timeout = ref 0 (* in seconds *)
+let max_memory = ref 0 (* in MB *)
 
 (*s: constant [[Main_semgrep_core.action]] *)
 (* action mode *)
@@ -296,7 +303,7 @@ let cache_computation file cache_file_of_file f =
       let file_cache = cache_file_of_file file in
       if Sys.file_exists file_cache && filemtime file_cache >= filemtime file
       then begin
-        if !verbose then pr2 ("using cache: " ^ file_cache);
+        if !Flag.debug then pr2 ("using cache: " ^ file_cache);
         let (version, file2, res) = Common2.get_value file_cache in
         if version <> Version.version
         then failwith (spf "Version mismatch! Clean the cache file %s"
@@ -345,7 +352,28 @@ let timeout_function lang = fun f ->
   in
   if timeout <= 0
   then f ()
-  else Common.timeout_function ~verbose:!verbose timeout f
+  else Common.timeout_function ~verbose:!Flag.debug timeout f
+
+(* from https://discuss.ocaml.org/t/todays-trick-memory-limits-with-gc-alarms/4431 *)
+let run_with_memory_limit limit_mb f =
+  if limit_mb = 0
+  then  f ()
+  else
+    let limit = limit_mb * 1024 * 1024 in
+    let limit_memory () =
+      let mem = (Gc.quick_stat ()).Gc.heap_words in
+      if mem > limit / (Sys.word_size / 8)
+      then begin
+          if !Flag.debug
+          then pr2 (spf "maxout allocated memory: %d" (mem*(Sys.word_size/8)));
+          raise Out_of_memory
+        end
+    in
+    let alarm = Gc.create_alarm limit_memory in
+    Fun.protect f ~finally:(fun () ->
+        Gc.delete_alarm alarm;
+        Gc.compact ()
+    )
 
 (*****************************************************************************)
 (* Parsing *)
@@ -369,7 +397,7 @@ let parse_generic lang file =
       file (Lang.string_of_lang lang) Version.version
      in
      cache_file_of_file full_filename)
- (fun () -> timeout_function lang (fun () ->
+ (fun () ->
  try
   (* finally calling the actual function *)
   let ast = Parse_code.parse_and_resolve_name_use_pfff_or_treesitter lang file
@@ -389,7 +417,7 @@ let parse_generic lang file =
   *  just Timeout for now.
   *)
  with Timeout -> Right Timeout
-  ))
+  )
   in
   match v with
   | Left ast -> ast
@@ -530,24 +558,47 @@ let get_final_files xs =
 let iter_generic_ast_of_files_and_get_matches_and_exn_to_errors f files =
   let matches_and_errors =
     files |> map (fun file ->
-       if !verbose then pr2 (spf "Analyzing %s" file);
+       if !Flag.debug then pr2 (spf "Analyzing %s" file);
+       let lang =
+          match Lang.lang_of_string_opt !lang with
+          | Some lang -> lang
+          | _ -> failwith (spf "no language specified")
+       in
+       if !Flag.debug then pr2 (spf "PARSING: %s" file);
        try
-         let lang =
-           match Lang.lang_of_string_opt !lang with
-            | Some lang -> lang
-            | _ ->
-               failwith (spf "no language specified")
-         in
-         if !debug then pr2 (spf "PARSING: %s" file);
-         let ast = parse_generic lang file in
+         run_with_memory_limit !max_memory (fun () ->
+         timeout_function lang (fun () ->
+          let ast = parse_generic lang file in
+          (* calling the hook *)
+          (f file lang ast, [])
 
-         (* calling the hook *)
-         f file lang ast, []
-
-       (* note that Error_code.exn_to_error now recognized Timeout
-        * and will generate a TimeoutError code for it
+           (* This is just to test -max_memory, to give
+            * a chance to the Gc.create_alarm to run even if the program does
+            * not even need to run the Gc. However, this has a slow perf
+            * penality on small programs, which is why it's better to keep
+            * it guarded when you're not testing -max_memory.
+            *)
+            |> (fun v -> (if !test then Gc.full_major()); v)
+         ))
+       with
+       (* note that Error_code.exn_to_error already handles Timeout
+        * and would generate a TimeoutError code for it, but we intercept
+        * Timeout here to give a better diagnostic.
         *)
-       with exn -> [], [Error_code.exn_to_error file exn]
+        | (Timeout | Out_of_memory) as exn ->
+            let str_opt =
+              match !Semgrep_generic.last_matched_rule with
+              | None -> None
+              | Some rule -> Some (spf " with ruleid %s" rule.R.id)
+            in
+            let loc = Parse_info.first_loc_of_file file in
+            [], [Error_code.mk_error_loc loc
+              (match exn with
+              | Timeout -> Error_code.Timeout str_opt
+              | Out_of_memory -> Error_code.OutOfMemory str_opt
+              | _ -> raise Impossible
+              )]
+        | exn -> [], [Error_code.exn_to_error file exn]
     )
   in
   let matches = matches_and_errors |> List.map fst |> List.flatten in
@@ -566,6 +617,8 @@ let print_matches_and_errors files matches errs =
      "stats", stats
   ] in
   let s = Json_io.string_of_json json in
+  if !Flag.debug
+  then pr2 ("returned JSON: "^ s);
   pr s
 (*e: function [[Main_semgrep_core.print_matches_and_errors]] *)
 
@@ -610,7 +663,7 @@ let semgrep_with_one_pattern xs =
 
   files |> List.iter (fun file ->
     (*s: [[Main_semgrep_core.semgrep_with_one_pattern()]] if [[verbose]] *)
-    if !verbose
+    if !Flag.debug
     then pr2 (spf "processing: %s" file);
     (*e: [[Main_semgrep_core.semgrep_with_one_pattern()]] if [[verbose]] *)
     let process file = sgrep_ast pattern file (create_ast file) in
@@ -643,7 +696,7 @@ let semgrep_with_one_pattern xs =
  *)
 let semgrep_with_rules rules_file xs =
   (*s: [[Main_semgrep_core.semgrep_with_rules()]] if [[verbose]] *)
-  if !verbose then pr2 (spf "Parsing %s" rules_file);
+  if !Flag.debug then pr2 (spf "Parsing %s" rules_file);
   (*e: [[Main_semgrep_core.semgrep_with_rules()]] if [[verbose]] *)
   let rules = Parse_rules.parse rules_file in
   let files = get_final_files xs in
@@ -679,7 +732,7 @@ module TR = Tainting_rule
 
 (*s: function [[Main_semgrep_core.tainting_with_rules]] *)
 let tainting_with_rules rules_file xs =
-  if !verbose then pr2 (spf "Parsing %s" rules_file);
+  if !Flag.debug then pr2 (spf "Parsing %s" rules_file);
   let rules = Parse_tainting_rules.parse rules_file in
 
   let files = get_final_files xs in
@@ -859,7 +912,7 @@ let all_actions () = [
 
   "-test_parse_lang", " <files or dirs>",
   Common.mk_action_n_arg
-    (Test_parsing.test_parse_lang !verbose !lang get_final_files);
+    (Test_parsing.test_parse_lang !Flag.debug !lang get_final_files);
  ]
 
 (*e: function [[Main_semgrep_core.all_actions]] *)
@@ -922,17 +975,22 @@ let options () =
     (*e: [[Main_semgrep_core.options]] other cases *)
     "-use_parsing_cache", Arg.Set_string use_parsing_cache,
     " <dir> save and use parsed ASTs in a cache at given directory. Caller responsiblity to clear cache";
-    "-verbose", Arg.Unit (fun () ->
-      verbose := true;
-      Flag_semgrep.verbose := true;
-    ),
-    " ";
-    "-debug", Arg.Set debug,
-    " add debugging information in the output (e.g., tracing)";
+    "-filter_irrelevant_rules", Arg.Set Flag.filter_irrelevant_rules,
+    " filter rules not containing any strings in target file";
+    "-no_filter_irrelevant_rules", Arg.Clear Flag.filter_irrelevant_rules,
+    " do not filter rules";
+    "-debug", Arg.Set Flag.debug,
+    " add debugging information in the output (tracing)";
+    "-debug_matching", Arg.Set Flag.debug_matching,
+    " add more debugging information on matching";
+    "-test", Arg.Set test,
+    " (internal) set test context";
     "-target_file", Arg.Set_string target_file,
     " <file> obtain list of targets to run patterns on";
     "-timeout", Arg.Set_int timeout,
-    " <int> timeout for parsing";
+    " <int> timeout for parsing (in seconds)";
+    "-max_memory", Arg.Set_int max_memory,
+    " <int> maximum memory (in MB)";
   ] @
   (*s: [[Main_semgrep_core.options]] concatenated flags *)
   Flag_parsing_cpp.cmdline_flags_macrofile () @
@@ -1006,7 +1064,11 @@ let main () =
   let argv =
    (Array.to_list Sys.argv) @
    (if Sys.getenv_opt "SEMGREP_CORE_DEBUG" <> None then ["-debug"] else[])@
-   (if Sys.getenv_opt "SEMGREP_CORE_PROFILE" <> None then ["-profile"] else[])
+   (if Sys.getenv_opt "SEMGREP_CORE_PROFILE" <> None then ["-profile"] else[])@
+   (match Sys.getenv_opt "SEMGREP_CORE_EXTRA" with
+   | Some s -> Common.split "[ \t]+" s
+   | None -> []
+   )
   in
 
   (* does side effect on many global flags *)
@@ -1018,7 +1080,7 @@ let main () =
   end
   in
 
-  if !debug then begin
+  if !Flag.debug then begin
     pr2 "Debug mode On";
     pr2 (spf "Executed as: %s" (Sys.argv|>Array.to_list|> String.concat " "));
   end;
@@ -1054,7 +1116,7 @@ let main () =
                semgrep_with_rules !rules_file (x::xs);
                if !profile then save_rules_file_in_tmp ();
             with exn -> begin
-             if !debug then save_rules_file_in_tmp ();
+             if !Flag.debug then save_rules_file_in_tmp ();
              pr (format_output_exception exn);
              exit 2
              end