+++ semgrep-core/tests/go/misc_tuple_order.go
@@ -1,0 +1,14 @@
+package foo
+
+func foo() {
+
+    if a, b := bar(); b {
+        fmt.Println("hello")
+    }
+
+    //ERROR: match
+    if a, b := bar(); a {
+        fmt.Println("hello")
+    }
+
+}